create a Fee and Package when editing a delivery

+++ db/migrate/20090421182220_create_packages.rb
@@ -2,10 +2,10 @@ class CreatePackages < ActiveRecord::Migration
   def self.up
     create_table :packages do |t|
       t.string :description
-      t.integer :weight_in_grams
-      t.integer :height_in_meters
-      t.integer :width_in_meters
-      t.integer :depth_in_meters
+      t.float :weight_in_grams
+      t.float :height_in_meters
+      t.float :width_in_meters
+      t.float :depth_in_meters
       t.timestamps
     end
   end
